+++ pycosio/_core/compat.py
@@ -63,7 +63,7 @@ if _py[0] == 2:
 
     # Missing "dir_fd" in "os" functions
 
-    def mkdir(path, mode=0o777, dir_fd=None):
+    def mkdir(path, mode=0o777, *, dir_fd=None):
         """
         Create a directory named path.
 
@@ -74,7 +74,7 @@ if _py[0] == 2:
         """
         _os.mkdir(path, mode)
 
-    def remove(path, dir_fd=None):
+    def remove(path, *, dir_fd=None):
         """
         Remove a file.
 
@@ -84,7 +84,7 @@ if _py[0] == 2:
         """
         _os.remove(path)
 
-    def rmdir(path, dir_fd=None):
+    def rmdir(path, *, dir_fd=None):
         """
         Remove a directory.
 
@@ -93,6 +93,27 @@ if _py[0] == 2:
             dir_fd: Ignored.
         """
         _os.rmdir(path)
+
+    def stat(path, *, dir_fd=None, follow_symlinks=True):
+        """
+        Get the status of a file.
+
+        Args:
+            path (str): Path.
+            dir_fd: Ignored.
+            follow_symlinks: Ignored
+        """
+        _os.stat(path)
+
+    def lstat(path, *, dir_fd=None):
+        """
+        Get the status of a file.
+
+        Args:
+            path (str): Path.
+            dir_fd: Ignored.
+        """
+        _os.lstat(path)
 
     # Missing "abc.ABC"
     ABC = _abc.ABCMeta('ABC', (object,), {})
@@ -114,6 +135,8 @@ else:
     mkdir = _os.mkdir
     remove = _os.remove
     rmdir = _os.rmdir
+    stat = _os.stat
+    lstat = _os.lstat
     copyfile = _shutil.copyfile
     ABC = _abc.ABC
     file_not_found_error = FileNotFoundError